// src/graphql-types.d.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  // tslint:disable-next-line:jsdoc-format
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
}
// ...
export interface Internal {
  content?: Maybe<Scalars["String"]>;
  contentDigest: Scalars["String"];
  description?: Maybe<Scalars["String"]>;
  fieldOwners?: Maybe<Array<Maybe<Scalars["String"]>>>;
  ignoreType?: Maybe<Scalars["Boolean"]>;
  mediaType?: Maybe<Scalars["String"]>;
  owner: Scalars["String"];
  type: Scalars["String"];
}
// ...
export interface MarkdownHeading {
  value?: Maybe<Scalars["String"]>;
  depth?: Maybe<Scalars["Int"]>;
}

export interface MarkdownHeadingFilterInput {
  value?: Maybe<StringQueryOperatorInput>;
  depth?: Maybe<IntQueryOperatorInput>;
}

export interface MarkdownHeadingFilterListInput {
  elemMatch?: Maybe<MarkdownHeadingFilterInput>;
}

export type MarkdownRemark = Node & {
  id: Scalars["ID"];
  parent?: Maybe<Node>;
  children: Node[];
  internal: Internal;
  frontmatter?: Maybe<MarkdownRemarkFrontmatter>;
  excerpt?: Maybe<Scalars["String"]>;
  canonical: Maybe<Scalars["String"]>;
  rawMarkdownBody?: Maybe<Scalars["String"]>;
  fileAbsolutePath?: Maybe<Scalars["String"]>;
  fields?: Maybe<MarkdownRemarkFields>;
  html?: Maybe<Scalars["String"]>;
  htmlAst?: Maybe<Scalars["JSON"]>;
  excerptAst?: Maybe<Scalars["JSON"]>;
  headings?: Maybe<Array<Maybe<MarkdownHeading>>>;
  timeToRead?: Maybe<Scalars["Int"]>;
  tableOfContents?: Maybe<Scalars["String"]>;
  wordCount?: Maybe<WordCount>;
};
// ...
export interface MarkdownRemarkFields {
  slug?: Maybe<Scalars["String"]>;
}
// ...
export interface MarkdownRemarkFrontmatter {
  title?: Maybe<Scalars["String"]>;
  date?: Maybe<Scalars["Date"]>;
  description?: Maybe<Scalars["String"]>;
  draft?: Maybe<Scalars["String"]>;
  canonical?: Maybe<Scalars["String"]>;
}
// ...
export interface Node {
  id: Scalars["ID"];
  parent?: Maybe<Node>;
  children: Node[];
  internal: Internal;
}
// ...
export interface WordCount {
  paragraphs?: Maybe<Scalars["Int"]>;
  sentences?: Maybe<Scalars["Int"]>;
  words?: Maybe<Scalars["Int"]>;
}